new toast for all sync, upload and code cleanup
